requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.10.6

# Configuration
django-environ==0.4.1
whitenoise==3.3.0


# Forms
django-braces==1.11.0
django-crispy-forms==1.6.1

# Models
django-model-utils==2.6.1

# Images
Pillow==4.0.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.31.0

# On Windows, you must download/install psycopg2 manually
# from http://www.lfd.uci.edu/~gohlke/pythonlibs/#psycopg


# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2016.10

# Redis support
django-redis==4.7.0
redis>=2.10.5



# Your custom requirements go here


defusedxml==0.4.1

djangorestframework==3.4.4
mysqlclient==1.3.10
oauthlib==1.1.2
python3-openid==3.0.10
requests==2.11.0
requests-oauthlib==0.6.2
python-memcached==1.58
git+https://github.com/mobolic/facebook-sdk.git#egg=facebook-sdk
celery==3.1.25
django-celery==3.1.17
